Increase the limit in the number of dimensions handled by CoordinateFormat from 32 to 64, for consistency with the limits used elswhere in the library.

+++ core/sis-referencing/src/main/java/org/apache/sis/geometry/CoordinateFormat.java
@@ -169,7 +169,7 @@ public class CoordinateFormat extends CompoundFormat<DirectPosition> {
      *
      * @see #negate(int)
      */
-    private transient int negate;
+    private transient long negate;
 
     /**
      * The time epochs. Non-null only if the at least on ordinate is to be formatted as a date.
@@ -355,17 +355,17 @@ public class CoordinateFormat extends CompoundFormat<DirectPosition> {
      * Remembers that ordinate values at the given dimension will need to have their sign reverted.
      */
     private void negate(final int dimension) {
-        if (dimension >= Integer.SIZE) {
+        if (dimension >= Long.SIZE) {
             throw new ArithmeticException(Errors.format(Errors.Keys.ExcessiveNumberOfDimensions_1, dimension));
         }
-        negate |= (1 << dimension);
+        negate |= (1L << dimension);
     }
 
     /**
      * Returns {@code true} if the value at the given dimension needs to have its sign reversed.
      */
     private boolean isNegative(final int dimension) {
-        return (dimension < Integer.SIZE) && (negate & (1 << dimension)) != 0;
+        return (dimension < Long.SIZE) && (negate & (1L << dimension)) != 0;
     }
 
     /**
